+++ automatedtests/src/androidTestR2_14_1/java/com/mux/stats/sdk/muxstats/automatedtests/ui/SimplePlayerTestActivity.java
@@ -11,6 +11,7 @@ import com.google.android.exoplayer2.DefaultRenderersFactory;
 import com.google.android.exoplayer2.MediaItem;
 import com.google.android.exoplayer2.MediaItem.AdsConfiguration;
 import com.google.android.exoplayer2.PlaybackPreparer;
+import com.google.android.exoplayer2.Player;
 import com.google.android.exoplayer2.RenderersFactory;
 import com.google.android.exoplayer2.SimpleExoPlayer;
 import com.google.android.exoplayer2.ext.ima.ImaAdsLoader;
@@ -35,7 +36,7 @@ import com.google.android.exoplayer2.util.Util;
 import com.mux.stats.sdk.muxstats.automatedtests.R;
 
 
-public class SimplePlayerTestActivity extends SimplePlayerBaseActivity implements PlaybackPreparer {
+public class SimplePlayerTestActivity extends SimplePlayerBaseActivity implements Player.Listener {
 
   MediaSourceFactory mediaSourceFactory;
 
@@ -65,12 +66,7 @@ public class SimplePlayerTestActivity extends SimplePlayerBaseActivity implement
             .setMediaSourceFactory(mediaSourceFactory)
             .setTrackSelector(trackSelector)
             .build();
-    playerView.setPlaybackPreparer(this);
-  }
-
-  @Override
-  public void preparePlayback() {
-    player.prepare();
+    player.addListener(this);
   }
 
   // This is for background playback, set appropriate notification and etc
@@ -198,5 +194,38 @@ public class SimplePlayerTestActivity extends SimplePlayerBaseActivity implement
       // TODO implement this
       Log.e(TAG, "onNotificationPosted");
     }
+  }
+
+  //////////////////////////////////////////////////////////////////////
+  ////// Player.EventListener //////////////////////////////////////////
+
+  @Override
+  public void onPlayerStateChanged(boolean playWhenReady, int playbackState) {
+    switch (playbackState) {
+      case Player.STATE_BUFFERING:
+        signalPlaybackBuffering();
+        break;
+      case Player.STATE_ENDED:
+        signalPlaybackEnded();
+        break;
+      case Player.STATE_READY:
+        // By the time we get here, it depends on playWhenReady to know if we're playing
+        if (playWhenReady) {
+          signalPlaybackStarted();
+        } else {
+          // TODO implement this
+//                    signalPlaybackPaused();
+        }
+      case Player.STATE_IDLE:
+        signalPlaybackStopped();
+        break;
+    }
+  }
+
+  @Override
+  public void onRepeatModeChanged(int repeatMode) {
+    activityLock.lock();
+    activityInitialized.signalAll();
+    activityLock.unlock();
   }
 }
