+++ automatedtests/src/androidTestR2_17_1/java/com/mux/stats/sdk/muxstats/automatedtests/ui/SimplePlayerTestActivity.java
@@ -39,7 +39,8 @@ import com.google.android.exoplayer2.util.Util;
 import com.mux.stats.sdk.muxstats.automatedtests.R;
 
 
-public class SimplePlayerTestActivity extends SimplePlayerBaseActivity {
+public class SimplePlayerTestActivity extends SimplePlayerBaseActivity implements
+    Player.Listener  {
 
   MediaSourceFactory mediaSourceFactory;
 
@@ -71,6 +72,7 @@ public class SimplePlayerTestActivity extends SimplePlayerBaseActivity {
             .setTrackSelector(trackSelector)
             .build();
     playerView.setPlayer(player);
+    player.addListener(this);
   }
 
   // This is for background playback, set appropriate notification and etc
@@ -203,4 +205,37 @@ public class SimplePlayerTestActivity extends SimplePlayerBaseActivity {
       Log.e(TAG, "onNotificationPosted");
     }
   }
+  //////////////////////////////////////////////////////////////////////
+  ////// Player.EventListener //////////////////////////////////////////
+
+  @Override
+  public void onPlayerStateChanged(boolean playWhenReady, int playbackState) {
+    switch (playbackState) {
+      case Player.STATE_BUFFERING:
+        signalPlaybackBuffering();
+        break;
+      case Player.STATE_ENDED:
+        signalPlaybackEnded();
+        break;
+      case Player.STATE_READY:
+        // By the time we get here, it depends on playWhenReady to know if we're playing
+        if (playWhenReady) {
+          signalPlaybackStarted();
+        } else {
+          // TODO implement this
+//                    signalPlaybackPaused();
+        }
+      case Player.STATE_IDLE:
+        signalPlaybackStopped();
+        break;
+    }
+  }
+
+  @Override
+  public void onRepeatModeChanged(int repeatMode) {
+    activityLock.lock();
+    activityInitialized.signalAll();
+    activityLock.unlock();
+  }
+
 }
