+++ automatedtests/src/androidTestR2_9_6/java/com/mux/stats/sdk/muxstats/automatedtests/ui/SimplePlayerTestActivity.java
@@ -7,6 +7,7 @@ import androidx.annotation.Nullable;
 import com.google.android.exoplayer2.C;
 import com.google.android.exoplayer2.DefaultRenderersFactory;
 import com.google.android.exoplayer2.ExoPlayerFactory;
+import com.google.android.exoplayer2.Player;
 import com.google.android.exoplayer2.RenderersFactory;
 import com.google.android.exoplayer2.SimpleExoPlayer;
 import com.google.android.exoplayer2.ext.ima.ImaAdsLoader;
@@ -28,7 +29,8 @@ import com.google.android.exoplayer2.util.Util;
 import com.mux.stats.sdk.muxstats.automatedtests.R;
 import java.lang.reflect.Constructor;
 
-public class SimplePlayerTestActivity extends SimplePlayerBaseActivity {
+public class SimplePlayerTestActivity extends SimplePlayerBaseActivity
+    implements Player.EventListener {
 
   public void initExoPlayer() {
     // Hopfully this will not channge the track selection set programmatically
@@ -44,6 +46,7 @@ public class SimplePlayerTestActivity extends SimplePlayerBaseActivity {
     trackSelector.setParameters(trackSelectorParameters);
     RenderersFactory renderersFactory = new DefaultRenderersFactory(/* context= */ this);
     player = ExoPlayerFactory.newSimpleInstance(this, renderersFactory, trackSelector);
+    player.addListener(this);
   }
 
   public void initAudioSession() {
@@ -145,5 +148,38 @@ public class SimplePlayerTestActivity extends SimplePlayerBaseActivity {
     player.setPlayWhenReady(playWhenReady);
     player.seekTo(playbackStartPosition);
     ((SimpleExoPlayer)player).prepare(testMediaSource, false, true);
+  }
+
+  //////////////////////////////////////////////////////////////////////
+  ////// Player.EventListener //////////////////////////////////////////
+
+  @Override
+  public void onPlayerStateChanged(boolean playWhenReady, int playbackState) {
+    switch (playbackState) {
+      case Player.STATE_BUFFERING:
+        signalPlaybackBuffering();
+        break;
+      case Player.STATE_ENDED:
+        signalPlaybackEnded();
+        break;
+      case Player.STATE_READY:
+        // By the time we get here, it depends on playWhenReady to know if we're playing
+        if (playWhenReady) {
+          signalPlaybackStarted();
+        } else {
+          // TODO implement this
+//                    signalPlaybackPaused();
+        }
+      case Player.STATE_IDLE:
+        signalPlaybackStopped();
+        break;
+    }
+  }
+
+  @Override
+  public void onRepeatModeChanged(int repeatMode) {
+    activityLock.lock();
+    activityInitialized.signalAll();
+    activityLock.unlock();
   }
 }
